# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AddTeamMemberRequest": {
        "properties": {
          "role_id": {
            "title": "Role Id",
            "type": "integer"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "role_id"
        ],
        "title": "AddTeamMemberRequest",
        "type": "object"
      },
...
      "BreakdownItem": {
        "properties": {
          "avg_resolution_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Resolution Hours"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "pending": {
            "title": "Pending",
            "type": "integer"
          },
          "sla_breached": {
            "title": "Sla Breached",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "key",
          "label",
          "total",
          "pending",
          "sla_breached",
          "avg_resolution_hours"
        ],
        "title": "BreakdownItem",
        "type": "object"
      },
...
      "CategoryOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_archived": {
            "default": false,
            "title": "Is Archived",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "CategoryOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RoleOut": {
        "properties": {
          "can_create_settings": {
            "title": "Can Create Settings",
            "type": "boolean"
          },
          "can_delete_settings": {
            "title": "Can Delete Settings",
            "type": "boolean"
          },
          "can_edit_settings": {
            "title": "Can Edit Settings",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_archived": {
            "title": "Is Archived",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "can_create_settings",
          "can_edit_settings",
          "can_delete_settings",
          "is_archived"
        ],
        "title": "RoleOut",
        "type": "object"
      },
...
      "TagOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_archived": {
            "default": false,
            "title": "Is Archived",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "TagOut",
        "type": "object"
      },
...
      "TeamChangeRequest": {
        "properties": {
          "team_id": {
            "title": "Team Id",
            "type": "integer"
          }
        },
        "required": [
          "team_id"
        ],
        "title": "TeamChangeRequest",
        "type": "object"
      },
      "TeamCreateRequest": {
        "properties": {
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "TeamCreateRequest",
        "type": "object"
      },
      "TeamDetailOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_default": {
            "default": false,
            "title": "Is Default",
            "type": "boolean"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/TeamMemberOut"
            },
            "title": "Members",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "members"
        ],
        "title": "TeamDetailOut",
        "type": "object"
      },
      "TeamMemberOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "role": {
            "$ref": "#/components/schemas/RoleOut"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "required": [
          "id",
          "user",
          "role"
        ],
        "title": "TeamMemberOut",
        "type": "object"
      },
      "TeamOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_default": {
            "default": false,
            "title": "Is Default",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "TeamOut",
        "type": "object"
      },
      "TeamUpdateRequest": {
        "properties": {
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "TeamUpdateRequest",
        "type": "object"
      },
...
      "TicketOut": {
        "properties": {
          "age_seconds": {
            "title": "Age Seconds",
            "type": "integer"
          },
          "business_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Id"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryOut"
          },
          "closed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "$ref": "#/components/schemas/UserOut"
          },
          "custom_field_values": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CustomFieldValueOut"
            },
            "title": "Custom Field Values",
            "type": "array"
          },
          "customer": {
            "title": "Customer",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "doctor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doctor Name"
          },
          "first_response_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Response At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "mobile_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mobile Number"
          },
          "owner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "priority": {
            "$ref": "#/components/schemas/TicketPriority"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "sla_breached": {
            "title": "Sla Breached",
            "type": "boolean"
          },
          "sla_due_at": {
            "format": "date-time",
            "title": "Sla Due At",
            "type": "string"
          },
          "sla_hours": {
            "title": "Sla Hours",
            "type": "integer"
          },
          "sla_remaining_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Remaining Seconds"
          },
          "slack_channel_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slack Channel Id"
          },
          "slack_message_ts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slack Message Ts"
          },
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "support_assignee": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "tags": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TagOut"
            },
            "title": "Tags",
            "type": "array"
          },
          "team": {
            "$ref": "#/components/schemas/TeamOut"
          },
          "ticket_number": {
            "title": "Ticket Number",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "ticket_number",
          "title",
          "description",
          "customer",
          "business_id",
          "mobile_number",
          "doctor_name",
          "category",
          "team",
          "priority",
          "status",
          "sla_hours",
          "sla_due_at",
          "owner",
          "support_assignee",
          "created_by",
          "slack_channel_id",
          "slack_message_ts",
          "created_at",
          "first_response_at",
          "resolved_at",
          "closed_at",
          "updated_at",
          "sla_breached",
          "sla_remaining_seconds",
          "age_seconds"
        ],
        "title": "TicketOut",
        "type": "object"
      },
      "TicketPriority": {
        "enum": [
          "low",
          "medium",
          "high",
          "urgent"
        ],
        "title": "TicketPriority",
        "type": "string"
      },
      "TicketStatus": {
        "enum": [
          "open",
          "in_progress",
          "pending",
          "resolved",
          "closed"
        ],
        "title": "TicketStatus",
        "type": "string"
      },
...
      "UpdateTeamMemberRoleRequest": {
        "properties": {
          "role_id": {
            "title": "Role Id",
            "type": "integer"
          }
        },
        "required": [
          "role_id"
        ],
        "title": "UpdateTeamMemberRoleRequest",
        "type": "object"
      },
      "UserOut": {
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slack_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slack User Id"
          }
        },
        "required": [
          "id",
          "email",
          "name"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Support Operations Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/analytics/breakdown/team": {
      "get": {
        "operationId": "breakdown_by_team_api_analytics_breakdown_team_get",
        "parameters": [
          {
            "in": "query",
            "name": "team_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Team Id"
            }
          },
          {
            "in": "query",
            "name": "owner_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          },
          {
            "in": "query",
            "name": "support_assignee_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Support Assignee Id"
            }
          },
          {
            "in": "query",
            "name": "category_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TicketPriority"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TicketStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "sla_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(breached|ok)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sla Status"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BreakdownItem"
                  },
                  "title": "Response Breakdown By Team Api Analytics Breakdown Team Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Breakdown By Team",
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/api/teams": {
      "get": {
        "operationId": "list_teams_api_teams_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TeamOut"
                  },
                  "title": "Response List Teams Api Teams Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Teams",
        "tags": [
          "lookups"
        ]
      },
      "post": {
        "operationId": "create_team_api_teams_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-dev-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Team",
        "tags": [
          "teams"
        ]
      }
    },
    "/api/teams/{team_id}": {
      "patch": {
        "operationId": "rename_team_api_teams__team_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "Team Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Team",
        "tags": [
          "teams"
        ]
      }
    },
    "/api/teams/{team_id}/detail": {
      "get": {
        "operationId": "get_team_detail_api_teams__team_id__detail_get",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "Team Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Team Detail",
        "tags": [
          "teams"
        ]
      }
    },
    "/api/teams/{team_id}/members": {
      "post": {
        "operationId": "add_member_api_teams__team_id__members_post",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "Team Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTeamMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMemberOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Member",
        "tags": [
          "teams"
        ]
      }
    },
    "/api/teams/{team_id}/members/{member_id}": {
      "delete": {
        "operationId": "remove_member_api_teams__team_id__members__member_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "Team Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "member_id",
            "required": true,
            "schema": {
              "title": "Member Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Email"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Member",
        "tags": [
          "teams"
        ]
      },
      "patch": {
        "operationId": "update_member_role_api_teams__team_id__members__member_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "Team Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "member_id",
            "required": true,
            "schema": {
              "title": "Member Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeamMemberRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMemberOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Member Role",
        "tags": [
          "teams"
        ]
      }
    },
    "/api/teams/{team_id}/set-default": {
      "post": {
        "operationId": "set_default_team_api_teams__team_id__set_default_post",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "Team Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Default Team",
        "tags": [
          "teams"
        ]
      }
    },
...
    "/api/tickets/{ticket_id}/team": {
      "post": {
        "operationId": "change_ticket_team_api_tickets__ticket_id__team_post",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "title": "Ticket Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Ticket Team",
        "tags": [
          "tickets"
        ]
      }
    },
...
  }
}